Extract modules, add a bit of doc
